fix: can not read `data` of undefined
This resolves an error in the `getData` function when the value of connected `context.provider.name` is empty.
This happens if the given provider name is not valid e.g. has a typo or something.
We can also think about throwing an error here with the message that the provider name was not initialized and has now the value undefined.